moved acceleratedMotion as PROGMEM

+++ Maze3D_on_ArduinoNano_src/Controller.h
@@ -1,4 +1,4 @@
 #pragma once
 
 void initController();
-void loopController(int32_t& x, int32_t& y, int32_t& angle, int around);
+void loopController(int32_t& x, int32_t& y, int16_t& angle, int around);
